split the e2e test to another test folder
